requirements.txt
flask>=3.0
paho-mqtt>=2.0
pyyaml>=6.0
